fix(syncer): Split code into different files
